Refactor action buttons in the icicle graph
