+++ ui/packages/shared/components/src/Skeletons/MetricsGraph.tsx
@@ -1,0 +1,128 @@
+// Copyright 2022 The Parca Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+const MetricsGraphSkeleton = () => (
+  <svg
+    fill="none"
+    height="452"
+    viewBox="0 0 1435 452"
+    width="100%"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <defs>
+      <linearGradient id="y-chart-shimmer" x1="0%" y1="0%" x2="0%" y2="100%">
+        <stop offset="0%" style={{stopColor: '#ebebeb', stopOpacity: 1}} />
+        <stop offset="50%" style={{stopColor: '#F6F6F6', stopOpacity: 1}}>
+          <animate
+            attributeName="offset"
+            values="-2; -2; 1"
+            keyTimes="0; 0.25; 1"
+            dur="2s"
+            repeatCount="indefinite"
+          ></animate>
+        </stop>
+        <stop offset="100%" style={{stopColor: '#ebebeb', stopOpacity: 1}} />
+      </linearGradient>
+      <linearGradient id="x-chart-shimmer" x1="0%" y1="0%" x2="100%" y2="0%">
+        <stop offset="0.599964" stop-color="#f3f3f3" stop-opacity="1">
+          <animate
+            attributeName="offset"
+            values="-2; -2; 1"
+            keyTimes="0; 0.25; 1"
+            dur="2s"
+            repeatCount="indefinite"
+          ></animate>
+        </stop>
+        <stop offset="1.59996" stop-color="#ecebeb" stop-opacity="1">
+          <animate
+            attributeName="offset"
+            values="-1; -1; 2"
+            keyTimes="0; 0.25; 1"
+            dur="2s"
+            repeatCount="indefinite"
+          ></animate>
+        </stop>
+        <stop offset="2.59996" stop-color="#f3f3f3" stop-opacity="1">
+          <animate
+            attributeName="offset"
+            values="0; 0; 3"
+            keyTimes="0; 0.25; 1"
+            dur="2s"
+            repeatCount="indefinite"
+          ></animate>
+        </stop>
+      </linearGradient>
+    </defs>
+
+    <path d="m3.5 146h19v111h-19z" fill="url(#y-chart-shimmer)" />
+    <g stroke="#ececec">
+      <path d="m53 19h1378v365h-1378z" />
+      <path d="m52.5 139.039h1379" />
+      <path d="m52.5 79.8652h1379" />
+      <path d="m52.5 198.213h1379" />
+      <path d="m52.5 257.387h1379" />
+      <path d="m52.5 316.561h1379" />
+      <path d="m284.412 18.5v366" />
+      <path d="m512.765 18.5v366" />
+      <path d="m739.322 18.5v366" />
+      <path d="m967.669 18.5v366" />
+      <path d="m1196.01 18.5v366" />
+    </g>
+    <path
+      id="line-chart"
+      d="m59.0469 219.018 15.3287 18.443 15.3521-9.261 15.3053-45.218 15.346 19.161 15.475 26.026 15.177-8.714 15.352 24.135 15.323-11.89 15.335-17.409 15.358 13.334 15.299-14.885 15.34 5.708 15.329 9.041 15.341 1.306 15.328-8.993 15.335 15.737 15.329 1.446 15.475-50.644 15.223 15.875 15.323 19.059 15.305-.505 15.335 18.037 15.34-10.759 15.329-13.82 15.335 19.318 15.404-20.67 15.3 4.73 15.305 6.21 15.352 5.957 15.323-10.731 15.487 23.078 15.171-15.997 15.328-41.612 15.341 19.894 15.328 28.445 15.323-11.888 15.358 18.418 15.317-15.533 15.335-5.416 15.358 14.761 15.311-17.545 15.358-.133 15.305 19.095 15.481-2.97 15.177-15.239 15.815 13.635 15.088 3.429 15.124-50.091 15.329 23.347 15.34 18.612 15.335-6.35 15.311 20.201 15.334-5.183 15.335-21.711 15.346 22.854 15.358-25.831 15.423 9.051 15.206 7.617 15.352 5.504 15.317-12.49 15.323 20.007 15.347-1.744 15.32-54.27 15.34 20.107 15.33 26.672 15.33-10.381 15.33 21.335 15.34-11.158 15.39-10.58 15.39 14.077 15.21-15.575 15.34 2.279 15.33 11.167 15.34-1.543 15.32-9.005 15.34 18.423 15.34-4.802 15.34-49.412 15.32 21.411 15.35 22.053 15.32-9.192 15.45 20.823 15.25-9.207 15.29-9.304 15.35 11.078 15.32-16.693 15.33 2.881 15.34 13.074 15.33 3.252"
+      stroke="#cecbcb"
+      stroke-width="1.00044"
+      // fill="url(#shimmerGradient)"
+    />
+    <path
+      id="line-chart"
+      d="m58.9473 245.545 15.3521-3.548 15.2995-10.784 15.4981-31.534 15.218 20.347 15.288 11.129 15.358 19.748 15.346 1.114 15.311-25.752 15.329 18.418 15.358-20.204 15.317 2.293 15.305 17.368 15.335-5.174 15.334-9.743 15.347 19.442 15.44.767 15.264-18.629 15.282-31.338 15.375 16.618 15.388 16.13 15.229 12.922 15.352 5.358 15.335-25.327 15.328 24.048 15.388-22.26 15.322-1.878 15.306 17.355 15.334-7.564 15.516-7.18 15.13 22.022 15.323-6.657 15.557-15.232 15.124-26.918 15.334 19.828 15.352 15.123 15.323 11 15.341 2.841 15.328-23.562 15.335 17.804 15.323-16.64 15.393-.623 15.428 18.277 15.218-5.88 15.276-5.341 15.733 5.15 14.995 5.253 15.317-16.748 15.334-22.767 15.352 10.862 15.37 14.608 15.417 14.536 15.147 7.289 15.393-22.514 15.411 14.514 15.293-23.282 15.253 7.477 15.528 14.062 15.246-6.945 15.376-6.239 15.194 19.758 15.311-3.22 15.343-15.686 15.37-27.035 15.36 11.388 15.29 19.319 15.31 14.414 15.49-1.095 15.19-18.602 15.35 19.75 15.31-19.083 15.44-1.349 15.24 11.252 15.33-2.912 15.34-8.1 15.35 15.276 15.82 2.633 14.84-10.049 15.41 13.715 15.33 5.569 15.46-34.366 15.08 36.473 15.38-11.922 15.35-33.985 15.39-2.682 15.34-30.202 15.23 21.728 15.33 43.202 15.31-3.283 15.51-16.943"
+      stroke="#cecbcb"
+      stroke-width="1.00044"
+    />
+    <path
+      id="line-chart"
+      d="m59.0469 376.57 15.3287.577 15.3521 3.072 15.3053-1.15 15.346-1.526 15.475 1.109 15.177-.92 15.352-1.553 15.323 1.931 15.335-2.873 15.358 3.053 15.299-.565 15.34.764 15.329.966 15.341-.194 15.328-1.92 15.335 1.92 15.329-2.471 15.475 3.415 15.223-1.141 15.323-.961 15.305 1.541h15.335l15.34-3.652 15.329.007 15.335 2.893 15.404-4.079 15.3 4.637 15.305-1.531 15.352-1.354 15.323.225 15.487 2.636 15.171-2.085 15.328.575 15.341.765 15.328-.959 15.323-.381 15.358-.966 15.317 2.502 15.335-1.152 15.358.567 15.311-1.519 15.358-2.902 15.305 6.159 15.481-2.543 15.177 1.235 15.815-1.854 15.088.398 15.124-.308 15.329.187 15.34.965 15.335-.201 15.311 1.354 15.334-1.347 15.335-.577 15.346-.185 15.358.005 15.423-.242 15.206-.532 15.352 2.106 15.317-3.077 15.323 3.082 15.347-.005 15.32-1.15 15.34 1.534 15.33-.384 15.33-2.494 15.33.381 15.34-.386 15.39-.206 15.39 1.325 15.21 1.383 15.34-.003 15.33-.769 15.34.959 15.32.196 15.34-.388 15.34-2.106 15.34 3.069 15.32-1.533 15.35.953 15.32-.747 15.45-.437 15.25.6 15.29.781 15.35-1.347 15.32.388 15.33.576 15.34-2.878 15.33.381"
+      stroke="#cecbcb"
+      stroke-width="1.00044"
+    />
+    <path
+      id="line-chart"
+      d="m59.0469 374.136 15.3287.577 15.3521 3.073 15.3053-1.151 15.346-1.526 15.475 1.109 15.177-.92 15.352-1.553 15.323 1.932 23.589-28.434 7.104 28.613 15.299-.565 15.34.764 15.329.966 15.341-.194 15.328-13.157 15.335 13.157 15.329-13.708 15.475 14.652 15.223-1.141 15.323-.961 15.305 1.541h15.335l15.34-3.652 15.329.008 15.335 2.892 15.404-4.079 15.3 4.637 15.305-1.531 15.352-1.354 36.469-11.019-5.659 13.88 15.171-2.084 15.328.575 15.341.764 15.328-.958 15.323-.381 15.358-.966 15.317 2.502 15.335-1.153 15.358-12.179 15.311 11.228 15.358-20.185h20.197 17.298 24.283l15.088 8.957 8.678 4.07 6.446 7.342 15.329.187 15.34.966 15.335-.202 15.311 1.354 15.335-1.346 15.334-.578 15.346-.184 15.358.005 15.423-.243 15.206-.532 15.352 2.107 53.597-21.903-22.957 21.908 15.347-.005 15.32-1.15 15.34 1.533 15.33-.383 15.33-2.495 11.64-6.381 19.03 6.376 15.39-.206 15.39 1.325 15.21 1.383 23.87-8.878 6.8 8.107 15.34.958 15.32.197 15.34-.389 15.34-2.106 15.34 3.07 12.13-13.907 18.54 13.327 15.32-.748 15.45-.436 15.25.599 30.64-68.106v67.541l15.32.388 15.33.575 15.34-2.878 15.33.381"
+      stroke="#cecbcb"
+      stroke-width="1.00044"
+    />
+    <path
+      id="line-chart"
+      d="m59.4253 311.341 15.3639-31.823 15.3111-8.503 15.3467 44.655 15.305-32.747 15.375 4.809 15.306 30.68 15.317-59.164 15.358 52.823 15.323-26.855 15.328 27.841 15.329-24.929 15.341-4.672 15.322 22.522 15.423-44.558 15.329 29.518 15.246.893 15.353 17.938 15.328-30.634 15.335 28.607 15.34-16.028 15.323-7.496 15.358 5.934 15.323-.148 15.323 5.552 15.323-.949 15.475 8.971 15.182-22.706 15.341-.095 15.328 24.82 15.335-23.647 15.329 7.546 15.34 1.228 15.358 24.572 15.305-99.189 15.388 102.745 15.276-34.185 15.34 35.483 15.329-54.983 15.446 18.656 15.217 25.016 15.335-17.115 15.358-29.763 15.34 57.72 15.294-19.619 15.346-5.962 15.335 4.516 15.364 14.094 15.317-33.665 15.352 24.837 15.352 7.248 15.276-21.143 15.329 6.236 15.346 15.096 15.323.386 15.44-35.985 15.223 32.561 15.335-17.486 15.328 24.829 15.353-21.43 15.317 13.483 15.334-13.638 15.339-5.36 15.33 12.625 15.36-15.465 15.32.92 15.32 27.657 15.33-32.297 15.36 36.252 15.31-20.852 15.33-7.122 15.34-11.095 15.33 33.376 15.32-24.929 15.34 17.625 15.33-41.721 15.35 22.521 15.44 14.635 15.21-10.153 15.35-12.878 15.33 36.424 15.38-20.165 15.3-10.066 15.34 11.725 15.32-10.225 15.33 7.146 15.32-1.029 15.38 18.261 15.3-20.819 15.33-27.061"
+      stroke="#cecbcb"
+      stroke-width="2.00088"
+    />
+    <path
+      id="line-chart"
+      d="m58.9473 259.676 15.3521 3.388 15.2995-10.393 15.4981 18.792 15.218 11.152 15.288-17.566 15.358 8.583 15.346-8.632 15.311 1.415 15.329 8.28 15.358-14.242 15.317-4.472 15.305 6.867 15.335 12.094 15.334-28.236 15.347 1.767 15.44 8.362 15.264 17.11 15.282-11.155 15.375 8.563 15.388 5.256 15.229-24.921 15.352-9.962 15.335 23.993 15.328-1.32 15.388-11.609 15.322 13.349 15.306-1.726 15.334-26.542 15.516 25.368 15.13 8.937 15.323-2.034 15.557-14.817 15.124 25.017 15.334-57.563 15.352 69.412 15.323-36.021 15.341 7.103 15.328 12.376 15.335-40.159 15.323 14.883 15.393 24.569 15.428-21.667 15.218 30.112 15.276-16.528 15.733 20.75 14.995-27.918 15.317 7.226 15.334-8.085 15.352 24.905 15.37-32.445 15.417 23.631 15.147-20.08 15.393-.187 15.411 38.217 15.293-30.714 15.253 12.839 15.528-29.234 15.246 31.882 15.376-30.976 15.194 27.814 15.311-19.879 15.343 24.747 15.37-31.188 15.36 15.579 15.29-.269 15.31-10.182 15.49 25.827 15.19-42.428 15.35 42.767 15.31-16.105 15.44 9.709 15.24-13.041 15.33 20.83 15.34-51.452 15.35 50.853 15.82-16.562 14.84-29.219 15.41 25.322 15.33-7.421 15.46 22.483 15.08-2.201 15.38-24.99 15.35 19.702 15.39-8.27 15.34 6.186 15.23-4.62 15.33 10.512 15.31-14.436 15.51-40.224"
+      stroke="#cecbcb"
+      stroke-width="1.00044"
+    />
+    <path
+      id="line-chart"
+      d="m59.4258 230.887 15.3638-21.474 15.3112 8.557 15.3462-18.572 15.305-6.134 15.376 10.155 15.305 9.855 15.317-7.108 15.358 3.269 15.323-16.506 15.329 22.153 15.329-33.546 15.34 31.983 15.323-8.265 15.422-1.92 15.329 16.405 15.247-28.742 15.352 20.399 15.329-28.741 15.334 26.511 15.341-11.969 15.323-.592 15.358 4.788 15.323 5.363 15.322 4.419 15.323-4.861 15.475 8.506 15.183-19.127 15.34 20.335 15.329-20.318 15.334 26.091 15.329-25.666 15.341 1.907 15.358-12.347 15.305-56.281 15.387 74.052 15.276 5.066 15.341 12.306 15.328-46.621 15.446 53.513 15.218-23.379 15.334 23.704 15.358-35.602 15.341 7.251 15.293-.099 15.347 8.942 15.334 20.743 15.364-42.153 15.317 17.498 15.352 9.68 15.352-.65 15.277 16.365 15.328-23.694 15.346-2.028 15.323-17.569 15.44 35.608 15.224-30.85 15.334 27.785 15.329-15.042 15.352-11.146 15.317 23.165 15.335 1.878 15.338-17.29 15.33 6.423 15.36 1.345 15.32-3.939 15.32-8.643 15.33 19.811 15.36-22.464 15.31 26.482 15.33-46.596 15.34 27.785 15.33-9.447 15.32 15.79 15.34 18.984 15.33-28.761 15.35 11.748 15.44-9.758 15.21 17.445 15.35-13.349 15.33-19.197 15.38 44.466 15.3-34.757 15.34 2.291 15.32 14.713 15.33-20.826 15.32 27.56 15.38-4.55 15.3 4.938 15.33-23.451"
+      stroke="#cecbcb"
+      stroke-width="1.00044"
+    />
+    <path id="x-shimmer-chart" d="m635 413.5h165v19h-165z" fill="url(#x-chart-shimmer)" />
+  </svg>
+);
+
+export default MetricsGraphSkeleton;
