+++ ui/packages/shared/components/src/Skeletons/Table.tsx
@@ -1,0 +1,423 @@
+// Copyright 2022 The Parca Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+const TableSkeleton = () => (
+  <svg
+    fill="none"
+    height="100%"
+    viewBox="0 0 1415 658"
+    width="100%"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <defs>
+      <linearGradient id="table-data" x1="0%" y1="0%" x2="100%" y2="0%">
+        <stop offset="0.599964" stop-color="#f3f3f3" stop-opacity="1">
+          <animate
+            attributeName="offset"
+            values="-2; -2; 1"
+            keyTimes="0; 0.25; 1"
+            dur="2s"
+            repeatCount="indefinite"
+          ></animate>
+        </stop>
+        <stop offset="1.59996" stop-color="#ecebeb" stop-opacity="1">
+          <animate
+            attributeName="offset"
+            values="-1; -1; 2"
+            keyTimes="0; 0.25; 1"
+            dur="2s"
+            repeatCount="indefinite"
+          ></animate>
+        </stop>
+        <stop offset="2.59996" stop-color="#f3f3f3" stop-opacity="1">
+          <animate
+            attributeName="offset"
+            values="0; 0; 3"
+            keyTimes="0; 0.25; 1"
+            dur="2s"
+            repeatCount="indefinite"
+          ></animate>
+        </stop>
+      </linearGradient>
+    </defs>
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="46.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="46.5" />
+    <path d="m1 64.5 1412 .0003" stroke="#e5e7eb" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="46.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="76.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="76.5" />
+    <path d="m1 94.5 1412 .0003" stroke="#e5e7eb" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="76.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="106.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="106.5" />
+    <path d="m1 124.5h1412" stroke="#e5e7eb" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="106.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="136.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="136.5" />
+    <path d="m1 154.5h1412" stroke="#e5e7eb" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="136.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="166.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="166.5" />
+    <path d="m1 184.5h1412" stroke="#e5e7eb" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="166.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="196.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="196.5" />
+    <path d="m1 214.5h1412" stroke="#e5e7eb" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="196.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="226.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="226.5" />
+    <path d="m1 244.5h1412" stroke="#e5e7eb" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="226.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="256.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="256.5" />
+    <path d="m1 274.5h1412" stroke="#e5e7eb" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="256.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="286.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="286.5" />
+    <path d="m1 304.5h1412" stroke="#e5e7eb" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="286.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="316.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="316.5" />
+    <path d="m1 334.5h1412" stroke="#e5e7eb" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="316.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="346.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="346.5" />
+    <path d="m1 364.5h1412" stroke="#e5e7eb" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="346.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="376.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="376.5" />
+    <path d="m1 394.5h1412" stroke="#e5e7eb" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="376.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="406.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="406.5" />
+    <path d="m1 424.5h1412" stroke="#e5e7eb" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="406.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="436.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="436.5" />
+    <path d="m1 454.5h1412" stroke="#e5e7eb" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="436.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="466.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="466.5" />
+    <path d="m1 484.5h1412" stroke="#e5e7eb" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="466.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="496.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="496.5" />
+    <path d="m1 514.5h1412" stroke="#e5e7eb" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="496.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="526.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="526.5" />
+    <path d="m1 544.5h1412" stroke="#e5e7eb" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="526.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="556.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="556.5" />
+    <path d="m1 574.5h1412" stroke="#e5e7eb" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="556.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="586.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="586.5" />
+    <path d="m1 604.5h1412" stroke="#e5e7eb" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="586.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="616.5" />
+    <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="616.5" />
+    <path d="m1 634.5h1412" stroke="#e5e7eb" />
+    <g fill="url(#table-data)" id="table-data">
+      <rect height="8" rx="4" width="400" x="268" y="616.5" />
+      <rect height="8" rx="4" width="39" x="18" y="646.5" />
+      <rect height="8" rx="4" width="71" x="110" y="646.5" />
+      <rect height="8" rx="4" width="400" x="268" y="646.5" />
+      <path d="m.5 0h1414v36h-1414z" fill="#faf9fb" />
+    </g>
+    <path
+      d="m27.0068 18.7617h-4.2109v4.2383h-1.6543v-9.9531h6.4668v1.3603h-4.8125v3.001h4.2109zm2.7754-6.2617h3.9854v9.1465h2.2422v1.3535h-6.2276v-1.3535h2.3379v-7.7861h-2.3379zm12.8174 10.5c-.0501-.0911-.0911-.2005-.123-.3281-.0274-.1322-.0524-.2689-.0752-.4102-.114.1185-.2438.2324-.3897.3418-.1413.1048-.3008.196-.4785.2735-.1777.082-.3691.1458-.5742.1914-.2051.0455-.4261.0683-.6631.0683-.3828 0-.7337-.0547-1.0527-.164-.3145-.114-.5834-.2689-.8067-.4649-.2278-.1914-.4056-.4215-.5332-.6904-.123-.2735-.1845-.5674-.1845-.8819 0-.4056.0797-.7633.2392-1.0732.1641-.3144.401-.5742.7109-.7793.2872-.1914.6403-.3372 1.0596-.4375.4238-.1003.9024-.1504 1.4356-.1504h1.1894v-.499c0-.1914-.0342-.3646-.1025-.5195-.0638-.155-.1572-.2894-.2803-.4034-.1322-.1139-.2939-.2005-.4853-.2597-.1869-.0638-.4034-.0957-.6495-.0957-.2233 0-.4192.0273-.5879.082-.164.0547-.3007.1299-.4101.2256-.1094.0866-.1914.1937-.2461.3213-.0501.123-.0752.2552-.0752.3965h-1.6406c.0045-.2963.0752-.5811.2119-.8545.1367-.2735.3372-.515.6016-.7246.2552-.2142.5696-.3829.9433-.5059.3783-.1276.8089-.1914 1.292-.1914.4329 0 .8363.0547 1.21.1641.3782.1048.7041.262.9775.4716.2734.2097.4876.4763.6426.7998.1549.3191.2324.6882.2324 1.1075v3.2881c0 .3326.0228.6334.0684.9023.0501.2643.1207.4922.2119.6836v.1162zm-1.9892-1.1758c.2141 0 .4124-.0273.5947-.082.1868-.0547.3554-.1253.5058-.2119s.2803-.1846.3897-.294c.1094-.1093.1937-.2233.2529-.3418v-1.3672h-1.0527c-.3555 0-.6631.0342-.9229.1026-.2597.0638-.4671.1572-.622.2803-.1368.1048-.2393.2301-.3077.3759-.0638.1459-.0957.3099-.0957.4922 0 .1504.0274.2917.0821.4238.0547.1277.1344.237.2392.3282.1048.0911.2347.164.3897.2187.1549.0502.3372.0752.5469.0752zm8.955-8.0254v1.8047h2.7823v1.2236h-2.7823v3.5342c0 .2735.0319.5036.0957.6905.0684.1822.1618.3258.2803.4306.1185.1094.2598.1869.4238.2324.1641.0456.3441.0684.5401.0684.1412 0 .2871-.0068.4375-.0205s.2962-.0319.4375-.0547c.1367-.0182.2643-.0387.3828-.0615.123-.0274.2256-.0501.3076-.0684l.1777 1.128c-.1185.0729-.262.1367-.4306.1914-.1641.0501-.3418.0934-.5332.1298-.1914.0365-.392.0638-.6016.0821-.2051.0228-.4101.0342-.6152.0342-.3737 0-.7178-.0502-1.0323-.1504-.3098-.1049-.5765-.2666-.7998-.4854-.2233-.2142-.3987-.4899-.5263-.8271-.1231-.3373-.1846-.7429-.1846-1.2168v-3.6368h-1.791v-1.2236h1.791v-1.8047z"
+      fill="#1f2937"
+    />
+    <path
+      d="m116.738 19.9375c-.045.4831-.159.9229-.342 1.3193-.182.392-.421.7269-.717 1.0049-.301.2826-.659.5013-1.074.6563-.41.1504-.865.2256-1.367.2256-.41 0-.788-.0547-1.134-.1641-.342-.1139-.648-.2734-.916-.4785-.26-.1914-.488-.4216-.684-.6905-.192-.2734-.353-.5719-.485-.8955-.133-.319-.233-.6653-.301-1.039-.064-.3737-.096-.7634-.096-1.169v-1.3603c0-.3874.032-.7611.096-1.1211s.159-.6996.287-1.0186c.137-.3554.312-.6767.526-.9638.219-.2917.47-.5401.752-.7452.269-.1868.565-.3326.889-.4375.328-.1048.686-.1572 1.073-.1572.524 0 .991.0775 1.401.2324.415.155.768.3737 1.06.6563.296.2871.529.6335.697 1.0391.173.4056.285.8567.335 1.3535h-1.64c-.028-.2963-.085-.5629-.171-.7998-.082-.2416-.196-.4466-.342-.6153-.15-.1686-.335-.2962-.554-.3828-.219-.0911-.481-.1367-.786-.1367-.214 0-.408.0296-.581.0889-.169.0592-.319.1435-.451.2529-.182.1322-.335.3053-.458.5195-.119.2097-.217.4444-.294.7041-.059.2279-.105.4717-.137.7315-.027.2552-.041.5172-.041.7861v1.374c0 .3327.018.6449.055.9366.041.2916.102.5582.184.7998.064.2005.144.3828.24.5468.1.1641.218.3054.355.4239.141.1276.303.2255.486.2939.186.0638.398.0957.635.0957.283 0 .533-.041.752-.123.219-.0821.406-.2028.561-.3623.15-.155.271-.3487.362-.5811s.15-.499.178-.7998zm6.576 3.0625-.095-1.0322c-.251.3691-.556.6585-.916.8681-.36.2051-.764.3077-1.21.3077-.383 0-.734-.0593-1.053-.1778-.314-.123-.586-.3122-.813-.5674-.224-.2552-.397-.581-.52-.9775-.123-.401-.185-.8796-.185-1.4355v-4.3819h1.634v4.3955c0 .3418.025.6289.075.8614.051.2278.13.4101.24.5468.109.1413.248.2416.417.3008.168.0593.367.0889.594.0889.237 0 .449-.0251.636-.0752.187-.0547.351-.1276.492-.2188.119-.082.224-.1777.315-.2871.091-.1139.168-.2392.232-.376v-5.2363h1.641v7.3965zm4.492-7.3965.048.8135c.141-.2962.334-.5287.581-.6973.25-.1686.551-.2529.902-.2529.182 0 .346.0205.492.0615s.271.1048.376.1914c.082.0638.155.1413.219.2324.064.0866.118.1869.164.3008.064-.1048.137-.2028.219-.2939.082-.0912.171-.1686.266-.2324.123-.0821.26-.1459.41-.1914.155-.0456.324-.0684.506-.0684.237 0 .454.041.65.123.2.0775.373.1983.519.3623.146.1732.258.3988.335.6768.082.2734.123.6038.123.9912v5.3799h-1.504v-5.3936c0-.1458-.016-.2688-.048-.3691-.031-.1048-.077-.1891-.136-.2529-.06-.0593-.132-.1026-.219-.1299-.082-.0274-.173-.041-.273-.041-.105 0-.196.0136-.274.041-.077.0273-.146.0661-.205.1162-.055.0456-.1.1003-.137.1641-.036.0592-.063.1276-.082.205v.0684.1025 5.4893h-1.476v-5.4004c0-.1458-.016-.2689-.048-.3691-.027-.1048-.071-.1869-.13-.2461-.059-.0593-.13-.1026-.212-.1299-.077-.0274-.166-.041-.267-.041-.109 0-.207.0159-.293.0478-.083.0274-.153.0707-.212.1299-.046.041-.087.0889-.123.1436-.032.0547-.062.1162-.089.1845v5.6807h-1.504v-7.3965zm12.325 7.3965-.096-1.0322c-.25.3691-.556.6585-.916.8681-.36.2051-.763.3077-1.21.3077-.383 0-.734-.0593-1.053-.1778-.314-.123-.585-.3122-.813-.5674-.223-.2552-.397-.581-.52-.9775-.123-.401-.184-.8796-.184-1.4355v-4.3819h1.634v4.3955c0 .3418.025.6289.075.8614.05.2278.13.4101.239.5468.109.1413.248.2416.417.3008.169.0593.367.0889.595.0889.237 0 .449-.0251.636-.0752.186-.0547.35-.1276.492-.2188.118-.082.223-.1777.314-.2871.091-.1139.169-.2392.233-.376v-5.2363h1.64v7.3965zm3.876-10.5h3.985v9.1465h2.242v1.3535h-6.227v-1.3535h2.338v-7.7861h-2.338zm12.817 10.5c-.05-.0911-.091-.2005-.123-.3281-.027-.1322-.052-.2689-.075-.4102-.114.1185-.244.2324-.39.3418-.141.1048-.3.196-.478.2735-.178.082-.369.1458-.574.1914-.205.0455-.427.0683-.663.0683-.383 0-.734-.0547-1.053-.164-.315-.114-.584-.2689-.807-.4649-.228-.1914-.405-.4215-.533-.6904-.123-.2735-.185-.5674-.185-.8819 0-.4056.08-.7633.24-1.0732.164-.3144.401-.5742.711-.7793.287-.1914.64-.3372 1.059-.4375.424-.1003.902-.1504 1.436-.1504h1.189v-.499c0-.1914-.034-.3646-.102-.5195-.064-.155-.158-.2894-.281-.4034-.132-.1139-.294-.2005-.485-.2597-.187-.0638-.403-.0957-.649-.0957-.224 0-.42.0273-.588.082-.164.0547-.301.1299-.411.2256-.109.0866-.191.1937-.246.3213-.05.123-.075.2552-.075.3965h-1.64c.004-.2963.075-.5811.211-.8545.137-.2735.338-.515.602-.7246.255-.2142.57-.3829.943-.5059.379-.1276.809-.1914 1.292-.1914.433 0 .837.0547 1.21.1641.379.1048.704.262.978.4716.273.2097.488.4763.642.7998.155.3191.233.6882.233 1.1075v3.2881c0 .3326.023.6334.068.9023.05.2643.121.4922.212.6836v.1162zm-1.989-1.1758c.214 0 .412-.0273.595-.082.187-.0547.355-.1253.506-.2119.15-.0866.28-.1846.389-.294.11-.1093.194-.2233.253-.3418v-1.3672h-1.053c-.355 0-.663.0342-.922.1026-.26.0638-.468.1572-.623.2803-.136.1048-.239.2301-.307.3759-.064.1459-.096.3099-.096.4922 0 .1504.027.2917.082.4238.055.1277.135.237.239.3282.105.0911.235.164.39.2187.155.0502.337.0752.547.0752zm8.955-8.0254v1.8047h2.782v1.2236h-2.782v3.5342c0 .2735.032.5036.096.6905.068.1822.162.3258.28.4306.119.1094.26.1869.424.2324.164.0456.344.0684.54.0684.141 0 .287-.0068.437-.0205.151-.0137.297-.0319.438-.0547.137-.0182.264-.0387.383-.0615.123-.0274.225-.0501.307-.0684l.178 1.128c-.118.0729-.262.1367-.431.1914-.164.0501-.341.0934-.533.1298-.191.0365-.392.0638-.601.0821-.205.0228-.411.0342-.616.0342-.373 0-.717-.0502-1.032-.1504-.31-.1049-.576-.2666-.8-.4854-.223-.2142-.398-.4899-.526-.8271-.123-.3373-.185-.7429-.185-1.2168v-3.6368h-1.791v-1.2236h1.791v-1.8047zm5.503 1.8047h3.924v6.043h2.174v1.3535h-6.098v-1.3535h2.276v-4.6826h-2.276zm2.126-1.914c0-.1322.023-.253.068-.3624.046-.1139.112-.2119.199-.2939.082-.0775.18-.1367.293-.1777.119-.0456.251-.0684.397-.0684.246 0 .444.0638.595.1914.155.1231.244.2826.266.4785.032.2006-.002.376-.102.5264-.1.1458-.246.2438-.438.2939-.095.0866-.209.1436-.342.1709-.132.0274-.278.016-.437-.0341-.164-.0502-.289-.139-.376-.2666-.082-.1322-.123-.2849-.123-.458zm8.326 9.3105-2.912-7.3965h1.695l1.832 5.3662.11.4922.109-.4922 1.812-5.3662h1.695l-2.898 7.3965zm9.441.1367c-.547 0-1.049-.0911-1.504-.2734-.452-.1823-.839-.4352-1.162-.7588-.324-.319-.575-.695-.752-1.1279-.174-.4375-.26-.9092-.26-1.4151v-.2803c0-.5787.091-1.1028.273-1.5722.183-.4694.433-.8705.752-1.2031.319-.3327.688-.5879 1.108-.7657.423-.1823.875-.2734 1.353-.2734.533 0 1.005.0911 1.415.2734.41.1778.754.4261 1.032.7452.278.3235.488.7086.629 1.1552.142.4466.212.9343.212 1.4629v.7041h-5.127v.0342c.05.3236.128.5947.233.8135.105.2187.248.4147.43.5879.183.1823.397.3235.643.4238.251.1003.524.1504.82.1504.406 0 .784-.0775 1.135-.2324.351-.1595.64-.3851.868-.6768l.875.8477c-.246.3554-.624.6744-1.135.957-.505.2825-1.118.4238-1.838.4238zm-.199-6.3437c-.232 0-.449.0433-.649.1299-.196.082-.372.2005-.526.3554-.155.1595-.285.3509-.39.5742s-.182.4763-.233.7588h3.48v-.1093c0-.2142-.039-.4262-.116-.6358-.078-.2142-.187-.401-.328-.5605-.137-.155-.31-.278-.52-.3692-.205-.0957-.444-.1435-.718-.1435z"
+      fill="#1f2937"
+    />
+    <path
+      d="m276.444 23h-1.647l-3.234-6.8496-.006 6.8496h-1.661v-9.9531h1.667l3.213 6.8222.014-6.8222h1.654zm6.747 0c-.05-.0911-.091-.2005-.123-.3281-.027-.1322-.052-.2689-.075-.4102-.114.1185-.244.2324-.389.3418-.142.1048-.301.196-.479.2735-.178.082-.369.1458-.574.1914-.205.0455-.426.0683-.663.0683-.383 0-.734-.0547-1.053-.164-.314-.114-.583-.2689-.807-.4649-.228-.1914-.405-.4215-.533-.6904-.123-.2735-.184-.5674-.184-.8819 0-.4056.079-.7633.239-1.0732.164-.3144.401-.5742.711-.7793.287-.1914.64-.3372 1.059-.4375.424-.1003.903-.1504 1.436-.1504h1.189v-.499c0-.1914-.034-.3646-.102-.5195-.064-.155-.157-.2894-.281-.4034-.132-.1139-.293-.2005-.485-.2597-.187-.0638-.403-.0957-.649-.0957-.224 0-.42.0273-.588.082-.164.0547-.301.1299-.41.2256-.11.0866-.192.1937-.246.3213-.051.123-.076.2552-.076.3965h-1.64c.004-.2963.075-.5811.212-.8545.136-.2735.337-.515.601-.7246.255-.2142.57-.3829.944-.5059.378-.1276.809-.1914 1.292-.1914.433 0 .836.0547 1.21.1641.378.1048.704.262.977.4716.274.2097.488.4763.643.7998.155.3191.232.6882.232 1.1075v3.2881c0 .3326.023.6334.068.9023.051.2643.121.4922.212.6836v.1162zm-1.989-1.1758c.214 0 .413-.0273.595-.082.187-.0547.355-.1253.506-.2119.15-.0866.28-.1846.389-.294.11-.1093.194-.2233.253-.3418v-1.3672h-1.052c-.356 0-.664.0342-.923.1026-.26.0638-.467.1572-.622.2803-.137.1048-.24.2301-.308.3759-.064.1459-.096.3099-.096.4922 0 .1504.028.2917.082.4238.055.1277.135.237.24.3282.104.0911.234.164.389.2187.155.0502.338.0752.547.0752zm6.604-6.2207.048.8135c.141-.2962.334-.5287.581-.6973.25-.1686.551-.2529.902-.2529.182 0 .346.0205.492.0615s.271.1048.376.1914c.082.0638.155.1413.219.2324.064.0866.118.1869.164.3008.064-.1048.137-.2028.219-.2939.082-.0912.171-.1686.266-.2324.123-.0821.26-.1459.41-.1914.155-.0456.324-.0684.506-.0684.237 0 .454.041.65.123.2.0775.373.1983.519.3623.146.1732.258.3988.335.6768.082.2734.123.6038.123.9912v5.3799h-1.504v-5.3936c0-.1458-.016-.2688-.048-.3691-.031-.1048-.077-.1891-.136-.2529-.06-.0593-.132-.1026-.219-.1299-.082-.0274-.173-.041-.273-.041-.105 0-.196.0136-.274.041-.077.0273-.146.0661-.205.1162-.055.0456-.1.1003-.137.1641-.036.0592-.063.1276-.082.205v.0684.1025 5.4893h-1.476v-5.4004c0-.1458-.016-.2689-.048-.3691-.027-.1048-.071-.1869-.13-.2461-.059-.0593-.13-.1026-.212-.1299-.077-.0274-.166-.041-.267-.041-.109 0-.207.0159-.293.0478-.083.0274-.153.0707-.212.1299-.046.041-.087.0889-.123.1436-.032.0547-.062.1162-.089.1845v5.6807h-1.504v-7.3965zm10.93 7.5332c-.547 0-1.048-.0911-1.504-.2734-.451-.1823-.838-.4352-1.162-.7588-.323-.319-.574-.695-.752-1.1279-.173-.4375-.259-.9092-.259-1.4151v-.2803c0-.5787.091-1.1028.273-1.5722s.433-.8705.752-1.2031c.319-.3327.688-.5879 1.107-.7657.424-.1823.875-.2734 1.354-.2734.533 0 1.005.0911 1.415.2734.41.1778.754.4261 1.032.7452.278.3235.488.7086.629 1.1552s.212.9343.212 1.4629v.7041h-5.127v.0342c.05.3236.128.5947.232.8135.105.2187.249.4147.431.5879.182.1823.397.3235.643.4238.25.1003.524.1504.82.1504.406 0 .784-.0775 1.135-.2324.351-.1595.64-.3851.868-.6768l.875.8477c-.246.3554-.624.6744-1.135.957-.506.2825-1.119.4238-1.839.4238zm-.198-6.3437c-.232 0-.449.0433-.649.1299-.196.082-.372.2005-.527.3554-.155.1595-.285.3509-.389.5742-.105.2233-.183.4763-.233.7588h3.48v-.1093c0-.2142-.039-.4262-.116-.6358-.078-.2142-.187-.401-.329-.5605-.136-.155-.31-.278-.519-.3692-.205-.0957-.444-.1435-.718-.1435z"
+      fill="#1f2937"
+    />
+  </svg>
+
+  // <svg
+  //   fill="none"
+  //   height="100%"
+  //   viewBox="0 0 1415 605"
+  //   width="100%"
+  //   xmlns="http://www.w3.org/2000/svg"
+  // >
+  //   <defs>
+  //     <linearGradient id="table-data" x1="0%" y1="0%" x2="100%" y2="0%">
+  //       <stop offset="0.599964" stop-color="#f3f3f3" stop-opacity="1">
+  //         <animate
+  //           attributeName="offset"
+  //           values="-2; -2; 1"
+  //           keyTimes="0; 0.25; 1"
+  //           dur="2s"
+  //           repeatCount="indefinite"
+  //         ></animate>
+  //       </stop>
+  //       <stop offset="1.59996" stop-color="#ecebeb" stop-opacity="1">
+  //         <animate
+  //           attributeName="offset"
+  //           values="-1; -1; 2"
+  //           keyTimes="0; 0.25; 1"
+  //           dur="2s"
+  //           repeatCount="indefinite"
+  //         ></animate>
+  //       </stop>
+  //       <stop offset="2.59996" stop-color="#f3f3f3" stop-opacity="1">
+  //         <animate
+  //           attributeName="offset"
+  //           values="0; 0; 3"
+  //           keyTimes="0; 0.25; 1"
+  //           dur="2s"
+  //           repeatCount="indefinite"
+  //         ></animate>
+  //       </stop>
+  //     </linearGradient>
+  //   </defs>
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="46.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="46.5" />
+  //   <path d="m1 64.5 1412 .0003" stroke="#e5e7eb" />
+  //   <g fill="url(#table-data)" id="table-data">
+  //     <rect height="8" rx="4" width="400" x="268" y="46.5" />
+  //     <rect height="8" rx="4" width="39" x="18" y="76.5" />
+  //     <rect height="8" rx="4" width="71" x="110" y="76.5" />
+  //   </g>
+  //   <path d="m1 94.5 1412 .0003" stroke="#e5e7eb" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="76.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="106.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="106.5" />
+  //   <path d="m1 124.5h1412" stroke="#e5e7eb" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="106.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="136.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="136.5" />
+  //   <path d="m1 154.5h1412" stroke="#e5e7eb" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="136.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="166.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="166.5" />
+  //   <path d="m1 184.5h1412" stroke="#e5e7eb" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="166.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="196.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="196.5" />
+  //   <path d="m1 214.5h1412" stroke="#e5e7eb" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="196.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="226.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="226.5" />
+  //   <path d="m1 244.5h1412" stroke="#e5e7eb" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="226.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="256.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="256.5" />
+  //   <path d="m1 274.5h1412" stroke="#e5e7eb" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="256.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="286.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="286.5" />
+  //   <path d="m1 304.5h1412" stroke="#e5e7eb" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="286.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="316.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="316.5" />
+  //   <path d="m1 334.5h1412" stroke="#e5e7eb" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="316.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="346.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="346.5" />
+  //   <path d="m1 364.5h1412" stroke="#e5e7eb" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="346.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="376.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="376.5" />
+  //   <path d="m1 394.5h1412" stroke="#e5e7eb" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="376.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="406.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="406.5" />
+  //   <path d="m1 424.5h1412" stroke="#e5e7eb" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="406.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="436.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="436.5" />
+  //   <path d="m1 454.5h1412" stroke="#e5e7eb" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="436.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="466.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="466.5" />
+  //   <path d="m1 484.5h1412" stroke="#e5e7eb" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="466.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="496.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="496.5" />
+  //   <path d="m1 514.5h1412" stroke="#e5e7eb" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="496.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="526.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="526.5" />
+  //   <path d="m1 544.5h1412" stroke="#e5e7eb" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="526.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="556.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="556.5" />
+  //   <path d="m1 574.5h1412" stroke="#e5e7eb" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="556.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="586.5" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="586.5" />
+  //   <path d="m1 604.5h1412" stroke="#e5e7eb" />
+  //   <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="586.5" />
+  //   <path d="m.5 0h1414v36h-1414z" fill="#faf9fb" />
+  //   <path
+  //     d="m21.0043 23v-10.1818h6.5227v1.5461h-4.6783v2.7643h4.2308v1.5461h-4.2308v4.3253zm9.9233-10.1818v10.1818h-1.7998v-10.1818zm4.0481 10.3359c-.4839 0-.9198-.0862-1.3076-.2585-.3844-.1757-.6893-.4342-.9147-.7756-.2221-.3414-.3331-.7623-.3331-1.2628 0-.4308.0795-.7871.2386-1.0688.1591-.2818.3762-.5072.6513-.6762s.585-.2966.9297-.3828c.348-.0895.7076-.1541 1.0788-.1939.4475-.0464.8104-.0878 1.0888-.1243.2784-.0398.4806-.0994.6065-.179.1293-.0828.1939-.2104.1939-.3828v-.0298c0-.3745-.111-.6645-.3331-.87-.222-.2055-.5419-.3083-.9595-.3083-.4408 0-.7905.0962-1.049.2884-.2552.1922-.4276.4193-.517.6811l-1.6804-.2386c.1325-.4641.3513-.8518.6562-1.1634.3049-.3149.6778-.5502 1.1186-.706.4408-.159.928-.2386 1.4617-.2386.3679 0 .7341.0431 1.0987.1293.3646.0861.6977.2287.9993.4275.3016.1956.5435.4624.7258.8004.1856.3381.2784.7607.2784 1.2678v5.1108h-1.7301v-1.049h-.0596c-.1094.2121-.2635.411-.4624.5966-.1955.1823-.4425.3298-.7408.4425-.2949.1093-.6413.164-1.039.164zm.4673-1.3224c.3613 0 .6745-.0713.9396-.2138.2652-.1458.469-.3381.6115-.5767.1459-.2386.2188-.4988.2188-.7805v-.8999c-.0564.0464-.1525.0895-.2884.1293-.1325.0397-.2817.0745-.4474.1044-.1657.0298-.3298.0563-.4922.0795s-.3033.0431-.4226.0597c-.2684.0364-.5087.0961-.7209.1789-.2121.0829-.3795.1989-.5021.3481-.1226.1458-.1839.3347-.1839.5667 0 .3315.1209.5817.3629.7507.2419.1691.5502.2536.9247.2536zm9.1875-6.4681v1.3921h-4.3899v-1.3921zm-3.3061-1.8295h1.7997v7.169c0 .242.0365.4276.1094.5568.0762.126.1757.2122.2983.2586s.2585.0696.4077.0696c.1126 0 .2154-.0083.3082-.0249.0961-.0166.169-.0315.2187-.0447l.3033 1.4069c-.0961.0332-.2337.0696-.4126.1094-.1757.0398-.3911.063-.6463.0696-.4508.0133-.8568-.0547-1.2181-.2038-.3612-.1525-.6479-.3878-.8601-.706-.2088-.3182-.3115-.7159-.3082-1.1932z"
+  //     fill="#000"
+  //   />
+  //   <path
+  //     d="m118.719 16.2536h-1.859c-.053-.305-.151-.5751-.293-.8104-.143-.2387-.32-.4408-.532-.6066-.212-.1657-.454-.29-.726-.3728-.269-.0862-.559-.1293-.87-.1293-.554 0-1.044.1392-1.472.4176-.427.2751-.762.6795-1.004 1.2131-.242.5303-.363 1.1783-.363 1.9439 0 .7789.121 1.4351.363 1.9687.245.5303.58.9314 1.004 1.2032.428.2684.917.4027 1.467.4027.305 0 .59-.0398.855-.1194.269-.0828.509-.2038.721-.3629.215-.1591.396-.3546.542-.5866.149-.232.252-.4972.308-.7955l1.859.01c-.069.4839-.22.9379-.452 1.3622-.229.4242-.529.7987-.9 1.1236-.371.3215-.805.5733-1.302.7556-.498.179-1.049.2685-1.656.2685-.895 0-1.694-.2071-2.396-.6214-.703-.4143-1.256-1.0126-1.661-1.7948-.404-.7822-.606-1.7202-.606-2.8139 0-1.0971.204-2.035.611-2.8139.408-.7822.963-1.3805 1.666-1.7948.702-.4143 1.498-.6214 2.386-.6214.567 0 1.094.0795 1.581.2386s.921.3928 1.303.701c.381.3049.694.6795.939 1.1236.249.4408.411.9446.487 1.5114zm6.477 3.5348v-4.4248h1.8v7.6364h-1.745v-1.3572h-.08c-.172.4275-.455.7772-.85 1.049-.391.2717-.873.4076-1.447.4076-.5 0-.942-.111-1.327-.3331-.381-.2253-.679-.5518-.895-.9794-.215-.4308-.323-.9512-.323-1.561v-4.8623h1.8v4.5838c0 .4839.132.8684.397 1.1535.266.285.614.4275 1.044.4275.266 0 .523-.0646.771-.1939.249-.1292.453-.3215.612-.5767.162-.2585.243-.5817.243-.9694zm3.651 3.2116v-7.6364h1.72v1.2976h.089c.159-.4375.423-.7789.791-1.0241.368-.2486.807-.3729 1.317-.3729.517 0 .953.126 1.308.3778.358.2486.61.5884.755 1.0192h.08c.169-.4242.454-.7623.855-1.0142.404-.2552.883-.3828 1.437-.3828.703 0 1.276.2221 1.72.6662s.666 1.0921.666 1.9439v5.1257h-1.804v-4.8473c0-.474-.126-.8203-.378-1.0391-.252-.222-.56-.3331-.925-.3331-.434 0-.774.1359-1.019.4077-.242.2685-.363.6181-.363 1.049v4.7628h-1.765v-4.9219c0-.3944-.119-.7093-.358-.9446-.235-.2353-.544-.353-.925-.353-.258 0-.494.0663-.706.1989-.212.1293-.381.3132-.507.5518-.126.2354-.189.5105-.189.8253v4.6435zm17.431-3.2116v-4.4248h1.8v7.6364h-1.745v-1.3572h-.08c-.172.4275-.455.7772-.85 1.049-.391.2717-.873.4076-1.447.4076-.5 0-.942-.111-1.327-.3331-.381-.2253-.679-.5518-.895-.9794-.215-.4308-.323-.9512-.323-1.561v-4.8623h1.8v4.5838c0 .4839.132.8684.397 1.1535.266.285.614.4275 1.045.4275.265 0 .522-.0646.77-.1939.249-.1292.453-.3215.612-.5767.162-.2585.243-.5817.243-.9694zm5.45-6.9702v10.1818h-1.799v-10.1818zm4.048 10.3359c-.483 0-.919-.0862-1.307-.2585-.385-.1757-.689-.4342-.915-.7756-.222-.3414-.333-.7623-.333-1.2628 0-.4308.08-.7871.239-1.0688.159-.2818.376-.5072.651-.6762s.585-.2966.93-.3828c.348-.0895.707-.1541 1.078-.1939.448-.0464.811-.0878 1.089-.1243.279-.0398.481-.0994.607-.179.129-.0828.194-.2104.194-.3828v-.0298c0-.3745-.111-.6645-.333-.87s-.542-.3083-.96-.3083c-.441 0-.79.0962-1.049.2884-.255.1922-.427.4193-.517.6811l-1.68-.2386c.132-.4641.351-.8518.656-1.1634.305-.3149.678-.5502 1.118-.706.441-.159.929-.2386 1.462-.2386.368 0 .734.0431 1.099.1293.364.0861.698.2287.999.4275.302.1956.544.4624.726.8004.186.3381.278.7607.278 1.2678v5.1108h-1.73v-1.049h-.059c-.11.2121-.264.411-.463.5966-.195.1823-.442.3298-.74.4425-.295.1093-.642.164-1.04.164zm.468-1.3224c.361 0 .674-.0713.939-.2138.266-.1458.469-.3381.612-.5767.146-.2386.219-.4988.219-.7805v-.8999c-.057.0464-.153.0895-.289.1293-.132.0397-.281.0745-.447.1044-.166.0298-.33.0563-.492.0795-.163.0232-.304.0431-.423.0597-.268.0364-.509.0961-.721.1789-.212.0829-.379.1989-.502.3481-.123.1458-.184.3347-.184.5667 0 .3315.121.5817.363.7507.242.1691.55.2536.925.2536zm9.187-6.4681v1.3921h-4.39v-1.3921zm-3.306-1.8295h1.8v7.169c0 .242.036.4276.109.5568.076.126.176.2122.299.2586.122.0464.258.0696.407.0696.113 0 .216-.0083.308-.0249.097-.0166.17-.0315.219-.0447l.303 1.4069c-.096.0332-.233.0696-.412.1094-.176.0398-.391.063-.646.0696-.451.0133-.857-.0547-1.218-.2038-.362-.1525-.648-.3878-.861-.706-.208-.3182-.311-.7159-.308-1.1932zm4.811 9.4659v-7.6364h1.8v7.6364zm.905-8.7202c-.285 0-.53-.0944-.736-.2834-.205-.1922-.308-.4225-.308-.691 0-.2718.103-.5021.308-.6911.206-.1922.451-.2883.736-.2883.289 0 .534.0961.736.2883.206.189.308.4193.308.6911 0 .2685-.102.4988-.308.691-.202.189-.447.2834-.736.2834zm9.567 1.0838-2.72 7.6364h-1.988l-2.72-7.6364h1.919l1.755 5.6726h.08l1.76-5.6726zm4.366 7.7855c-.765 0-1.427-.159-1.984-.4772-.553-.3215-.979-.7756-1.277-1.3622-.299-.59-.448-1.2844-.448-2.0831 0-.7855.149-1.4749.448-2.0682.301-.5966.722-1.0606 1.263-1.3921.54-.3347 1.175-.5021 1.904-.5021.47 0 .914.0762 1.332.2287.421.1491.792.3812 1.114.696.325.3149.58.7159.765 1.2031.186.4839.279 1.0607.279 1.7302v.5518h-6.259v-1.2131h4.534c-.004-.3447-.078-.6512-.224-.9197-.146-.2718-.35-.4856-.612-.6413-.258-.1558-.56-.2337-.904-.2337-.368 0-.692.0895-.97.2685-.278.1756-.495.4076-.651.696-.153.285-.231.5982-.234.9396v1.059c0 .4441.081.8252.244 1.1434.162.3149.389.5568.681.7259.292.1657.633.2486 1.024.2486.262 0 .499-.0365.711-.1094.212-.0762.396-.1873.552-.3331s.273-.3265.353-.5419l1.68.1889c-.106.4441-.308.8319-.606 1.1634-.295.3281-.673.5833-1.134.7656-.461.179-.988.2684-1.581.2684z"
+  //     fill="#000"
+  //   />
+  //   <path
+  //     d="m278.371 12.8182v10.1818h-1.64l-4.798-6.9354h-.084v6.9354h-1.845v-10.1818h1.651l4.792 6.9403h.09v-6.9403zm4.128 10.3359c-.484 0-.92-.0862-1.307-.2585-.385-.1757-.69-.4342-.915-.7756-.222-.3414-.333-.7623-.333-1.2628 0-.4308.079-.7871.238-1.0688.159-.2818.377-.5072.652-.6762s.585-.2966.929-.3828c.348-.0895.708-.1541 1.079-.1939.448-.0464.811-.0878 1.089-.1243.278-.0398.481-.0994.606-.179.13-.0828.194-.2104.194-.3828v-.0298c0-.3745-.111-.6645-.333-.87s-.542-.3083-.959-.3083c-.441 0-.791.0962-1.049.2884-.255.1922-.428.4193-.517.6811l-1.681-.2386c.133-.4641.352-.8518.657-1.1634.304-.3149.677-.5502 1.118-.706.441-.159.928-.2386 1.462-.2386.368 0 .734.0431 1.099.1293.364.0861.697.2287.999.4275.301.1956.543.4624.726.8004.185.3381.278.7607.278 1.2678v5.1108h-1.73v-1.049h-.06c-.109.2121-.263.411-.462.5966-.196.1823-.443.3298-.741.4425-.295.1093-.641.164-1.039.164zm.467-1.3224c.362 0 .675-.0713.94-.2138.265-.1458.469-.3381.612-.5767.145-.2386.218-.4988.218-.7805v-.8999c-.056.0464-.152.0895-.288.1293-.133.0397-.282.0745-.447.1044-.166.0298-.33.0563-.493.0795-.162.0232-.303.0431-.422.0597-.269.0364-.509.0961-.721.1789-.212.0829-.38.1989-.502.3481-.123.1458-.184.3347-.184.5667 0 .3315.121.5817.363.7507.242.1691.55.2536.924.2536zm5.375 1.1683v-7.6364h1.72v1.2976h.089c.159-.4375.423-.7789.791-1.0241.368-.2486.807-.3729 1.317-.3729.517 0 .953.126 1.308.3778.358.2486.61.5884.756 1.0192h.079c.169-.4242.454-.7623.855-1.0142.405-.2552.884-.3828 1.437-.3828.703 0 1.276.2221 1.72.6662s.666 1.0921.666 1.9439v5.1257h-1.804v-4.8473c0-.474-.126-.8203-.378-1.0391-.252-.222-.56-.3331-.925-.3331-.434 0-.774.1359-1.019.4077-.242.2685-.363.6181-.363 1.049v4.7628h-1.765v-4.9219c0-.3944-.119-.7093-.358-.9446-.235-.2353-.543-.353-.925-.353-.258 0-.493.0663-.706.1989-.212.1293-.381.3132-.507.5518-.126.2354-.189.5105-.189.8253v4.6435zm15.945.1491c-.766 0-1.427-.159-1.984-.4772-.553-.3215-.979-.7756-1.277-1.3622-.299-.59-.448-1.2844-.448-2.0831 0-.7855.149-1.4749.448-2.0682.301-.5966.722-1.0606 1.262-1.3921.541-.3347 1.175-.5021 1.904-.5021.471 0 .915.0762 1.333.2287.421.1491.792.3812 1.113.696.325.3149.58.7159.766 1.2031.186.4839.278 1.0607.278 1.7302v.5518h-6.259v-1.2131h4.534c-.003-.3447-.078-.6512-.223-.9197-.146-.2718-.35-.4856-.612-.6413-.258-.1558-.56-.2337-.905-.2337-.368 0-.691.0895-.969.2685-.279.1756-.496.4076-.651.696-.153.285-.231.5982-.234.9396v1.059c0 .4441.081.8252.243 1.1434.163.3149.39.5568.682.7259.291.1657.633.2486 1.024.2486.262 0 .499-.0365.711-.1094.212-.0762.396-.1873.552-.3331.155-.1458.273-.3265.352-.5419l1.681.1889c-.106.4441-.308.8319-.607 1.1634-.295.3281-.672.5833-1.133.7656-.461.179-.988.2684-1.581.2684z"
+  //     fill="#000"
+  //   />
+  // </svg>
+);
+
+export default TableSkeleton;
+
+<svg
+  fill="none"
+  height="100%"
+  viewBox="0 0 1415 658"
+  width="100%"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <defs>
+    <linearGradient id="table-data" x1="0%" y1="0%" x2="100%" y2="0%">
+      <stop offset="0.599964" stop-color="#f3f3f3" stop-opacity="1">
+        <animate
+          attributeName="offset"
+          values="-2; -2; 1"
+          keyTimes="0; 0.25; 1"
+          dur="2s"
+          repeatCount="indefinite"
+        ></animate>
+      </stop>
+      <stop offset="1.59996" stop-color="#ecebeb" stop-opacity="1">
+        <animate
+          attributeName="offset"
+          values="-1; -1; 2"
+          keyTimes="0; 0.25; 1"
+          dur="2s"
+          repeatCount="indefinite"
+        ></animate>
+      </stop>
+      <stop offset="2.59996" stop-color="#f3f3f3" stop-opacity="1">
+        <animate
+          attributeName="offset"
+          values="0; 0; 3"
+          keyTimes="0; 0.25; 1"
+          dur="2s"
+          repeatCount="indefinite"
+        ></animate>
+      </stop>
+    </linearGradient>
+  </defs>
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="46.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="46.5" />
+  <path d="m1 64.5 1412 .0003" stroke="#e5e7eb" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="46.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="76.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="76.5" />
+  <path d="m1 94.5 1412 .0003" stroke="#e5e7eb" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="76.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="106.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="106.5" />
+  <path d="m1 124.5h1412" stroke="#e5e7eb" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="106.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="136.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="136.5" />
+  <path d="m1 154.5h1412" stroke="#e5e7eb" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="136.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="166.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="166.5" />
+  <path d="m1 184.5h1412" stroke="#e5e7eb" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="166.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="196.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="196.5" />
+  <path d="m1 214.5h1412" stroke="#e5e7eb" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="196.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="226.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="226.5" />
+  <path d="m1 244.5h1412" stroke="#e5e7eb" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="226.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="256.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="256.5" />
+  <path d="m1 274.5h1412" stroke="#e5e7eb" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="256.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="286.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="286.5" />
+  <path d="m1 304.5h1412" stroke="#e5e7eb" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="286.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="316.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="316.5" />
+  <path d="m1 334.5h1412" stroke="#e5e7eb" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="316.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="346.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="346.5" />
+  <path d="m1 364.5h1412" stroke="#e5e7eb" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="346.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="376.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="376.5" />
+  <path d="m1 394.5h1412" stroke="#e5e7eb" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="376.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="406.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="406.5" />
+  <path d="m1 424.5h1412" stroke="#e5e7eb" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="406.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="436.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="436.5" />
+  <path d="m1 454.5h1412" stroke="#e5e7eb" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="436.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="466.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="466.5" />
+  <path d="m1 484.5h1412" stroke="#e5e7eb" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="466.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="496.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="496.5" />
+  <path d="m1 514.5h1412" stroke="#e5e7eb" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="496.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="526.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="526.5" />
+  <path d="m1 544.5h1412" stroke="#e5e7eb" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="526.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="556.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="556.5" />
+  <path d="m1 574.5h1412" stroke="#e5e7eb" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="556.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="586.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="586.5" />
+  <path d="m1 604.5h1412" stroke="#e5e7eb" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="400" x="268" y="586.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="39" x="18" y="616.5" />
+  <rect fill="url(#table-data)" id="table-data" height="8" rx="4" width="71" x="110" y="616.5" />
+  <path d="m1 634.5h1412" stroke="#e5e7eb" />
+  <g fill="url(#table-data)" id="table-data">
+    <rect height="8" rx="4" width="400" x="268" y="616.5" />
+    <rect height="8" rx="4" width="39" x="18" y="646.5" />
+    <rect height="8" rx="4" width="71" x="110" y="646.5" />
+    <rect height="8" rx="4" width="400" x="268" y="646.5" />
+    <path d="m.5 0h1414v36h-1414z" fill="#faf9fb" />
+  </g>
+  <path
+    d="m27.0068 18.7617h-4.2109v4.2383h-1.6543v-9.9531h6.4668v1.3603h-4.8125v3.001h4.2109zm2.7754-6.2617h3.9854v9.1465h2.2422v1.3535h-6.2276v-1.3535h2.3379v-7.7861h-2.3379zm12.8174 10.5c-.0501-.0911-.0911-.2005-.123-.3281-.0274-.1322-.0524-.2689-.0752-.4102-.114.1185-.2438.2324-.3897.3418-.1413.1048-.3008.196-.4785.2735-.1777.082-.3691.1458-.5742.1914-.2051.0455-.4261.0683-.6631.0683-.3828 0-.7337-.0547-1.0527-.164-.3145-.114-.5834-.2689-.8067-.4649-.2278-.1914-.4056-.4215-.5332-.6904-.123-.2735-.1845-.5674-.1845-.8819 0-.4056.0797-.7633.2392-1.0732.1641-.3144.401-.5742.7109-.7793.2872-.1914.6403-.3372 1.0596-.4375.4238-.1003.9024-.1504 1.4356-.1504h1.1894v-.499c0-.1914-.0342-.3646-.1025-.5195-.0638-.155-.1572-.2894-.2803-.4034-.1322-.1139-.2939-.2005-.4853-.2597-.1869-.0638-.4034-.0957-.6495-.0957-.2233 0-.4192.0273-.5879.082-.164.0547-.3007.1299-.4101.2256-.1094.0866-.1914.1937-.2461.3213-.0501.123-.0752.2552-.0752.3965h-1.6406c.0045-.2963.0752-.5811.2119-.8545.1367-.2735.3372-.515.6016-.7246.2552-.2142.5696-.3829.9433-.5059.3783-.1276.8089-.1914 1.292-.1914.4329 0 .8363.0547 1.21.1641.3782.1048.7041.262.9775.4716.2734.2097.4876.4763.6426.7998.1549.3191.2324.6882.2324 1.1075v3.2881c0 .3326.0228.6334.0684.9023.0501.2643.1207.4922.2119.6836v.1162zm-1.9892-1.1758c.2141 0 .4124-.0273.5947-.082.1868-.0547.3554-.1253.5058-.2119s.2803-.1846.3897-.294c.1094-.1093.1937-.2233.2529-.3418v-1.3672h-1.0527c-.3555 0-.6631.0342-.9229.1026-.2597.0638-.4671.1572-.622.2803-.1368.1048-.2393.2301-.3077.3759-.0638.1459-.0957.3099-.0957.4922 0 .1504.0274.2917.0821.4238.0547.1277.1344.237.2392.3282.1048.0911.2347.164.3897.2187.1549.0502.3372.0752.5469.0752zm8.955-8.0254v1.8047h2.7823v1.2236h-2.7823v3.5342c0 .2735.0319.5036.0957.6905.0684.1822.1618.3258.2803.4306.1185.1094.2598.1869.4238.2324.1641.0456.3441.0684.5401.0684.1412 0 .2871-.0068.4375-.0205s.2962-.0319.4375-.0547c.1367-.0182.2643-.0387.3828-.0615.123-.0274.2256-.0501.3076-.0684l.1777 1.128c-.1185.0729-.262.1367-.4306.1914-.1641.0501-.3418.0934-.5332.1298-.1914.0365-.392.0638-.6016.0821-.2051.0228-.4101.0342-.6152.0342-.3737 0-.7178-.0502-1.0323-.1504-.3098-.1049-.5765-.2666-.7998-.4854-.2233-.2142-.3987-.4899-.5263-.8271-.1231-.3373-.1846-.7429-.1846-1.2168v-3.6368h-1.791v-1.2236h1.791v-1.8047z"
+    fill="#1f2937"
+  />
+  <path
+    d="m116.738 19.9375c-.045.4831-.159.9229-.342 1.3193-.182.392-.421.7269-.717 1.0049-.301.2826-.659.5013-1.074.6563-.41.1504-.865.2256-1.367.2256-.41 0-.788-.0547-1.134-.1641-.342-.1139-.648-.2734-.916-.4785-.26-.1914-.488-.4216-.684-.6905-.192-.2734-.353-.5719-.485-.8955-.133-.319-.233-.6653-.301-1.039-.064-.3737-.096-.7634-.096-1.169v-1.3603c0-.3874.032-.7611.096-1.1211s.159-.6996.287-1.0186c.137-.3554.312-.6767.526-.9638.219-.2917.47-.5401.752-.7452.269-.1868.565-.3326.889-.4375.328-.1048.686-.1572 1.073-.1572.524 0 .991.0775 1.401.2324.415.155.768.3737 1.06.6563.296.2871.529.6335.697 1.0391.173.4056.285.8567.335 1.3535h-1.64c-.028-.2963-.085-.5629-.171-.7998-.082-.2416-.196-.4466-.342-.6153-.15-.1686-.335-.2962-.554-.3828-.219-.0911-.481-.1367-.786-.1367-.214 0-.408.0296-.581.0889-.169.0592-.319.1435-.451.2529-.182.1322-.335.3053-.458.5195-.119.2097-.217.4444-.294.7041-.059.2279-.105.4717-.137.7315-.027.2552-.041.5172-.041.7861v1.374c0 .3327.018.6449.055.9366.041.2916.102.5582.184.7998.064.2005.144.3828.24.5468.1.1641.218.3054.355.4239.141.1276.303.2255.486.2939.186.0638.398.0957.635.0957.283 0 .533-.041.752-.123.219-.0821.406-.2028.561-.3623.15-.155.271-.3487.362-.5811s.15-.499.178-.7998zm6.576 3.0625-.095-1.0322c-.251.3691-.556.6585-.916.8681-.36.2051-.764.3077-1.21.3077-.383 0-.734-.0593-1.053-.1778-.314-.123-.586-.3122-.813-.5674-.224-.2552-.397-.581-.52-.9775-.123-.401-.185-.8796-.185-1.4355v-4.3819h1.634v4.3955c0 .3418.025.6289.075.8614.051.2278.13.4101.24.5468.109.1413.248.2416.417.3008.168.0593.367.0889.594.0889.237 0 .449-.0251.636-.0752.187-.0547.351-.1276.492-.2188.119-.082.224-.1777.315-.2871.091-.1139.168-.2392.232-.376v-5.2363h1.641v7.3965zm4.492-7.3965.048.8135c.141-.2962.334-.5287.581-.6973.25-.1686.551-.2529.902-.2529.182 0 .346.0205.492.0615s.271.1048.376.1914c.082.0638.155.1413.219.2324.064.0866.118.1869.164.3008.064-.1048.137-.2028.219-.2939.082-.0912.171-.1686.266-.2324.123-.0821.26-.1459.41-.1914.155-.0456.324-.0684.506-.0684.237 0 .454.041.65.123.2.0775.373.1983.519.3623.146.1732.258.3988.335.6768.082.2734.123.6038.123.9912v5.3799h-1.504v-5.3936c0-.1458-.016-.2688-.048-.3691-.031-.1048-.077-.1891-.136-.2529-.06-.0593-.132-.1026-.219-.1299-.082-.0274-.173-.041-.273-.041-.105 0-.196.0136-.274.041-.077.0273-.146.0661-.205.1162-.055.0456-.1.1003-.137.1641-.036.0592-.063.1276-.082.205v.0684.1025 5.4893h-1.476v-5.4004c0-.1458-.016-.2689-.048-.3691-.027-.1048-.071-.1869-.13-.2461-.059-.0593-.13-.1026-.212-.1299-.077-.0274-.166-.041-.267-.041-.109 0-.207.0159-.293.0478-.083.0274-.153.0707-.212.1299-.046.041-.087.0889-.123.1436-.032.0547-.062.1162-.089.1845v5.6807h-1.504v-7.3965zm12.325 7.3965-.096-1.0322c-.25.3691-.556.6585-.916.8681-.36.2051-.763.3077-1.21.3077-.383 0-.734-.0593-1.053-.1778-.314-.123-.585-.3122-.813-.5674-.223-.2552-.397-.581-.52-.9775-.123-.401-.184-.8796-.184-1.4355v-4.3819h1.634v4.3955c0 .3418.025.6289.075.8614.05.2278.13.4101.239.5468.109.1413.248.2416.417.3008.169.0593.367.0889.595.0889.237 0 .449-.0251.636-.0752.186-.0547.35-.1276.492-.2188.118-.082.223-.1777.314-.2871.091-.1139.169-.2392.233-.376v-5.2363h1.64v7.3965zm3.876-10.5h3.985v9.1465h2.242v1.3535h-6.227v-1.3535h2.338v-7.7861h-2.338zm12.817 10.5c-.05-.0911-.091-.2005-.123-.3281-.027-.1322-.052-.2689-.075-.4102-.114.1185-.244.2324-.39.3418-.141.1048-.3.196-.478.2735-.178.082-.369.1458-.574.1914-.205.0455-.427.0683-.663.0683-.383 0-.734-.0547-1.053-.164-.315-.114-.584-.2689-.807-.4649-.228-.1914-.405-.4215-.533-.6904-.123-.2735-.185-.5674-.185-.8819 0-.4056.08-.7633.24-1.0732.164-.3144.401-.5742.711-.7793.287-.1914.64-.3372 1.059-.4375.424-.1003.902-.1504 1.436-.1504h1.189v-.499c0-.1914-.034-.3646-.102-.5195-.064-.155-.158-.2894-.281-.4034-.132-.1139-.294-.2005-.485-.2597-.187-.0638-.403-.0957-.649-.0957-.224 0-.42.0273-.588.082-.164.0547-.301.1299-.411.2256-.109.0866-.191.1937-.246.3213-.05.123-.075.2552-.075.3965h-1.64c.004-.2963.075-.5811.211-.8545.137-.2735.338-.515.602-.7246.255-.2142.57-.3829.943-.5059.379-.1276.809-.1914 1.292-.1914.433 0 .837.0547 1.21.1641.379.1048.704.262.978.4716.273.2097.488.4763.642.7998.155.3191.233.6882.233 1.1075v3.2881c0 .3326.023.6334.068.9023.05.2643.121.4922.212.6836v.1162zm-1.989-1.1758c.214 0 .412-.0273.595-.082.187-.0547.355-.1253.506-.2119.15-.0866.28-.1846.389-.294.11-.1093.194-.2233.253-.3418v-1.3672h-1.053c-.355 0-.663.0342-.922.1026-.26.0638-.468.1572-.623.2803-.136.1048-.239.2301-.307.3759-.064.1459-.096.3099-.096.4922 0 .1504.027.2917.082.4238.055.1277.135.237.239.3282.105.0911.235.164.39.2187.155.0502.337.0752.547.0752zm8.955-8.0254v1.8047h2.782v1.2236h-2.782v3.5342c0 .2735.032.5036.096.6905.068.1822.162.3258.28.4306.119.1094.26.1869.424.2324.164.0456.344.0684.54.0684.141 0 .287-.0068.437-.0205.151-.0137.297-.0319.438-.0547.137-.0182.264-.0387.383-.0615.123-.0274.225-.0501.307-.0684l.178 1.128c-.118.0729-.262.1367-.431.1914-.164.0501-.341.0934-.533.1298-.191.0365-.392.0638-.601.0821-.205.0228-.411.0342-.616.0342-.373 0-.717-.0502-1.032-.1504-.31-.1049-.576-.2666-.8-.4854-.223-.2142-.398-.4899-.526-.8271-.123-.3373-.185-.7429-.185-1.2168v-3.6368h-1.791v-1.2236h1.791v-1.8047zm5.503 1.8047h3.924v6.043h2.174v1.3535h-6.098v-1.3535h2.276v-4.6826h-2.276zm2.126-1.914c0-.1322.023-.253.068-.3624.046-.1139.112-.2119.199-.2939.082-.0775.18-.1367.293-.1777.119-.0456.251-.0684.397-.0684.246 0 .444.0638.595.1914.155.1231.244.2826.266.4785.032.2006-.002.376-.102.5264-.1.1458-.246.2438-.438.2939-.095.0866-.209.1436-.342.1709-.132.0274-.278.016-.437-.0341-.164-.0502-.289-.139-.376-.2666-.082-.1322-.123-.2849-.123-.458zm8.326 9.3105-2.912-7.3965h1.695l1.832 5.3662.11.4922.109-.4922 1.812-5.3662h1.695l-2.898 7.3965zm9.441.1367c-.547 0-1.049-.0911-1.504-.2734-.452-.1823-.839-.4352-1.162-.7588-.324-.319-.575-.695-.752-1.1279-.174-.4375-.26-.9092-.26-1.4151v-.2803c0-.5787.091-1.1028.273-1.5722.183-.4694.433-.8705.752-1.2031.319-.3327.688-.5879 1.108-.7657.423-.1823.875-.2734 1.353-.2734.533 0 1.005.0911 1.415.2734.41.1778.754.4261 1.032.7452.278.3235.488.7086.629 1.1552.142.4466.212.9343.212 1.4629v.7041h-5.127v.0342c.05.3236.128.5947.233.8135.105.2187.248.4147.43.5879.183.1823.397.3235.643.4238.251.1003.524.1504.82.1504.406 0 .784-.0775 1.135-.2324.351-.1595.64-.3851.868-.6768l.875.8477c-.246.3554-.624.6744-1.135.957-.505.2825-1.118.4238-1.838.4238zm-.199-6.3437c-.232 0-.449.0433-.649.1299-.196.082-.372.2005-.526.3554-.155.1595-.285.3509-.39.5742s-.182.4763-.233.7588h3.48v-.1093c0-.2142-.039-.4262-.116-.6358-.078-.2142-.187-.401-.328-.5605-.137-.155-.31-.278-.52-.3692-.205-.0957-.444-.1435-.718-.1435z"
+    fill="#1f2937"
+  />
+  <path
+    d="m276.444 23h-1.647l-3.234-6.8496-.006 6.8496h-1.661v-9.9531h1.667l3.213 6.8222.014-6.8222h1.654zm6.747 0c-.05-.0911-.091-.2005-.123-.3281-.027-.1322-.052-.2689-.075-.4102-.114.1185-.244.2324-.389.3418-.142.1048-.301.196-.479.2735-.178.082-.369.1458-.574.1914-.205.0455-.426.0683-.663.0683-.383 0-.734-.0547-1.053-.164-.314-.114-.583-.2689-.807-.4649-.228-.1914-.405-.4215-.533-.6904-.123-.2735-.184-.5674-.184-.8819 0-.4056.079-.7633.239-1.0732.164-.3144.401-.5742.711-.7793.287-.1914.64-.3372 1.059-.4375.424-.1003.903-.1504 1.436-.1504h1.189v-.499c0-.1914-.034-.3646-.102-.5195-.064-.155-.157-.2894-.281-.4034-.132-.1139-.293-.2005-.485-.2597-.187-.0638-.403-.0957-.649-.0957-.224 0-.42.0273-.588.082-.164.0547-.301.1299-.41.2256-.11.0866-.192.1937-.246.3213-.051.123-.076.2552-.076.3965h-1.64c.004-.2963.075-.5811.212-.8545.136-.2735.337-.515.601-.7246.255-.2142.57-.3829.944-.5059.378-.1276.809-.1914 1.292-.1914.433 0 .836.0547 1.21.1641.378.1048.704.262.977.4716.274.2097.488.4763.643.7998.155.3191.232.6882.232 1.1075v3.2881c0 .3326.023.6334.068.9023.051.2643.121.4922.212.6836v.1162zm-1.989-1.1758c.214 0 .413-.0273.595-.082.187-.0547.355-.1253.506-.2119.15-.0866.28-.1846.389-.294.11-.1093.194-.2233.253-.3418v-1.3672h-1.052c-.356 0-.664.0342-.923.1026-.26.0638-.467.1572-.622.2803-.137.1048-.24.2301-.308.3759-.064.1459-.096.3099-.096.4922 0 .1504.028.2917.082.4238.055.1277.135.237.24.3282.104.0911.234.164.389.2187.155.0502.338.0752.547.0752zm6.604-6.2207.048.8135c.141-.2962.334-.5287.581-.6973.25-.1686.551-.2529.902-.2529.182 0 .346.0205.492.0615s.271.1048.376.1914c.082.0638.155.1413.219.2324.064.0866.118.1869.164.3008.064-.1048.137-.2028.219-.2939.082-.0912.171-.1686.266-.2324.123-.0821.26-.1459.41-.1914.155-.0456.324-.0684.506-.0684.237 0 .454.041.65.123.2.0775.373.1983.519.3623.146.1732.258.3988.335.6768.082.2734.123.6038.123.9912v5.3799h-1.504v-5.3936c0-.1458-.016-.2688-.048-.3691-.031-.1048-.077-.1891-.136-.2529-.06-.0593-.132-.1026-.219-.1299-.082-.0274-.173-.041-.273-.041-.105 0-.196.0136-.274.041-.077.0273-.146.0661-.205.1162-.055.0456-.1.1003-.137.1641-.036.0592-.063.1276-.082.205v.0684.1025 5.4893h-1.476v-5.4004c0-.1458-.016-.2689-.048-.3691-.027-.1048-.071-.1869-.13-.2461-.059-.0593-.13-.1026-.212-.1299-.077-.0274-.166-.041-.267-.041-.109 0-.207.0159-.293.0478-.083.0274-.153.0707-.212.1299-.046.041-.087.0889-.123.1436-.032.0547-.062.1162-.089.1845v5.6807h-1.504v-7.3965zm10.93 7.5332c-.547 0-1.048-.0911-1.504-.2734-.451-.1823-.838-.4352-1.162-.7588-.323-.319-.574-.695-.752-1.1279-.173-.4375-.259-.9092-.259-1.4151v-.2803c0-.5787.091-1.1028.273-1.5722s.433-.8705.752-1.2031c.319-.3327.688-.5879 1.107-.7657.424-.1823.875-.2734 1.354-.2734.533 0 1.005.0911 1.415.2734.41.1778.754.4261 1.032.7452.278.3235.488.7086.629 1.1552s.212.9343.212 1.4629v.7041h-5.127v.0342c.05.3236.128.5947.232.8135.105.2187.249.4147.431.5879.182.1823.397.3235.643.4238.25.1003.524.1504.82.1504.406 0 .784-.0775 1.135-.2324.351-.1595.64-.3851.868-.6768l.875.8477c-.246.3554-.624.6744-1.135.957-.506.2825-1.119.4238-1.839.4238zm-.198-6.3437c-.232 0-.449.0433-.649.1299-.196.082-.372.2005-.527.3554-.155.1595-.285.3509-.389.5742-.105.2233-.183.4763-.233.7588h3.48v-.1093c0-.2142-.039-.4262-.116-.6358-.078-.2142-.187-.401-.329-.5605-.136-.155-.31-.278-.519-.3692-.205-.0957-.444-.1435-.718-.1435z"
+    fill="#1f2937"
+  />
+</svg>;
