+++ ui/packages/shared/profile/src/ProfileIcicleGraph/index.tsx
@@ -13,27 +13,21 @@
 
 import React, {Fragment, useCallback, useEffect, useMemo} from 'react';
 
-import {Menu, Transition} from '@headlessui/react';
 import {Icon} from '@iconify/react';
 import {AnimatePresence, motion} from 'framer-motion';
 
 import {Flamegraph, FlamegraphArrow} from '@parca/client';
-import {Button, IS, Select, useParcaContext, useURLState} from '@parca/components';
+import {Button, IcicleGraphSkeleton, Select, useParcaContext, useURLState} from '@parca/components';
 import {USER_PREFERENCES, useUserPreference} from '@parca/hooks';
 import {capitalizeOnlyFirstLetter, divide, type NavigateFunction} from '@parca/utilities';
 
 import {useProfileViewContext} from '../ProfileView/ProfileViewContext';
 import DiffLegend from '../components/DiffLegend';
+import GroupByDropdown from './ActionButtons/GroupByDropdown';
+import RuntimeFilterDropdown from './ActionButtons/RuntimeFilterDropdown';
+import SortBySelect from './ActionButtons/SortBySelect';
 import IcicleGraph from './IcicleGraph';
-import IcicleGraphArrow, {
-  FIELD_CUMULATIVE,
-  FIELD_DIFF,
-  FIELD_FUNCTION_FILE_NAME,
-  FIELD_FUNCTION_NAME,
-  FIELD_LABELS,
-  FIELD_LOCATION_ADDRESS,
-  FIELD_MAPPING_FILE,
-} from './IcicleGraphArrow';
+import IcicleGraphArrow, {FIELD_FUNCTION_NAME} from './IcicleGraphArrow';
 
 const numberFormatter = new Intl.NumberFormat('en-US');
 
@@ -83,7 +77,7 @@ const ShowHideLegendButton = ({navigateTo}: {navigateTo?: NavigateFunction}): JS
     <>
       {colorProfileName === 'default' || compareMode ? null : (
         <Button
-          className="gap-2"
+          className="gap-2 w-max"
           variant="neutral"
           onClick={() => setColorStackLegend(isColorStackLegendEnabled ? 'false' : 'true')}
         >
@@ -174,110 +168,6 @@ const GroupAndSortActionButtons = ({navigateTo}: {navigateTo?: NavigateFunction}
   );
 };
 
-const RuntimeToggle = ({
-  id,
-  state,
-  toggle,
-  label,
-  description,
-}: {
-  id: string;
-  state: boolean;
-  toggle: () => void;
-  label: string;
-  description: string;
-}): JSX.Element => {
-  return (
-    <div key={id} className="relative flex items-start">
-      <div className="flex h-6 items-center">
-        <input
-          id={id}
-          name={id}
-          type="checkbox"
-          className="h-4 w-4 rounded border-gray-300 text-indigo-600 focus:ring-indigo-600"
-          checked={state}
-          onChange={() => toggle()}
-        />
-      </div>
-      <div className="ml-3 text-sm leading-6">
-        <label htmlFor={id} className="font-medium text-gray-900 dark:text-gray-200">
-          {label}
-        </label>
-        <p className="text-gray-500 dark:text-gray-400">{description}</p>
-      </div>
-    </div>
-  );
-};
-
-const RuntimeFilterDropdown = ({
-  showRuntimeRuby,
-  toggleShowRuntimeRuby,
-  showRuntimePython,
-  toggleShowRuntimePython,
-  showInterpretedOnly,
-  toggleShowInterpretedOnly,
-}: {
-  showRuntimeRuby: boolean;
-  toggleShowRuntimeRuby: () => void;
-  showRuntimePython: boolean;
-  toggleShowRuntimePython: () => void;
-  showInterpretedOnly: boolean;
-  toggleShowInterpretedOnly: () => void;
-}): React.JSX.Element => {
-  return (
-    <div>
-      <label className="text-sm">Runtimes</label>
-      <Menu as="div" className="relative text-left">
-        <div>
-          <Menu.Button className="relative w-full cursor-default rounded-md border bg-white py-2 pl-3 pr-10 text-left text-sm shadow-sm focus:border-indigo-500 focus:outline-none focus:ring-1 focus:ring-indigo-500 dark:border-gray-600 dark:bg-gray-900 sm:text-sm">
-            <span className="block overflow-x-hidden text-ellipsis">Runtimes</span>
-            <span className="pointer-events-none absolute inset-y-0 right-0 ml-3 flex items-center pr-2 text-gray-400">
-              <Icon icon="heroicons:chevron-down-20-solid" aria-hidden="true" />
-            </span>
-          </Menu.Button>
-        </div>
-
-        <Transition
-          as={Fragment}
-          leave="transition ease-in duration-100"
-          leaveFrom="opacity-100"
-          leaveTo="opacity-0"
-        >
-          <Menu.Items className="absolute left-0 z-10 mt-1 min-w-[400px] overflow-auto rounded-md bg-gray-50 py-1 text-base shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none dark:border-gray-600 dark:bg-gray-900 dark:ring-white dark:ring-opacity-20 sm:text-sm">
-            <div className="p-4">
-              <fieldset>
-                <div className="space-y-5">
-                  <RuntimeToggle
-                    id="show-runtime-ruby"
-                    state={showRuntimeRuby}
-                    toggle={toggleShowRuntimeRuby}
-                    label="Ruby"
-                    description="Show Ruby runtime functions."
-                  />
-                  <RuntimeToggle
-                    id="show-runtime-python"
-                    state={showRuntimePython}
-                    toggle={toggleShowRuntimePython}
-                    label="Python"
-                    description="Show Python runtime functions."
-                  />
-                  <RuntimeToggle
-                    id="show-interpreted-only"
-                    state={showInterpretedOnly}
-                    toggle={toggleShowInterpretedOnly}
-                    label="Interpreted Only"
-                    description="Show only interpreted functions."
-                  />
-                </div>
-              </fieldset>
-            </div>
-          </Menu.Items>
-        </Transition>
-      </Menu>
-    </div>
-  );
-};
-
 const ProfileIcicleGraph = function ProfileIcicleGraphNonMemo({
   graph,
   arrow,
@@ -292,7 +182,7 @@ const ProfileIcicleGraph = function ProfileIcicleGraphNonMemo({
   error,
   width,
 }: ProfileIcicleGraphProps): JSX.Element {
-  const {loader, onError, authenticationErrorMessage} = useParcaContext();
+  const {onError, authenticationErrorMessage} = useParcaContext();
   const {compareMode} = useProfileViewContext();
 
   const [storeSortBy = FIELD_FUNCTION_NAME] = useURLState({
@@ -342,6 +232,7 @@ const ProfileIcicleGraph = function ProfileIcicleGraphNonMemo({
           <ShowHideLegendButton navigateTo={navigateTo} />
           <Button
             variant="neutral"
+            className="w-max"
             onClick={() => setNewCurPath([])}
             disabled={curPath.length === 0}
           >
@@ -353,7 +244,7 @@ const ProfileIcicleGraph = function ProfileIcicleGraphNonMemo({
   }, [navigateTo, arrow, curPath, setNewCurPath, setActionButtons, loading]);
 
   if (loading) {
-    return <div className="h-auto">{<IS />}</div>;
+    return <div className="h-auto">{<IcicleGraphSkeleton />}</div>;
   }
 
   if (error != null) {
@@ -426,171 +317,6 @@ const ProfileIcicleGraph = function ProfileIcicleGraphNonMemo({
         </p>
       </motion.div>
     </AnimatePresence>
-  );
-};
-
-const groupByOptions = [
-  {
-    value: FIELD_FUNCTION_NAME,
-    label: 'Function Name',
-    description: 'Stacktraces are grouped by function names.',
-    disabled: true,
-  },
-  {
-    value: FIELD_LABELS,
-    label: 'Labels',
-    description: 'Stacktraces are grouped by pprof labels.',
-    disabled: false,
-  },
-  {
-    value: FIELD_FUNCTION_FILE_NAME,
-    label: 'Filename',
-    description: 'Stacktraces are grouped by filenames.',
-    disabled: false,
-  },
-  {
-    value: FIELD_LOCATION_ADDRESS,
-    label: 'Address',
-    description: 'Stacktraces are grouped by addresses.',
-    disabled: false,
-  },
-  {
-    value: FIELD_MAPPING_FILE,
-    label: 'Binary',
-    description: 'Stacktraces are grouped by binaries.',
-    disabled: false,
-  },
-];
-
-const GroupByDropdown = ({
-  groupBy,
-  toggleGroupBy,
-}: {
-  groupBy: string[];
-  toggleGroupBy: (key: string) => void;
-}): React.JSX.Element => {
-  const label =
-    groupBy.length === 0
-      ? 'Nothing'
-      : groupBy.length === 1
-      ? groupByOptions.find(option => option.value === groupBy[0])?.label
-      : 'Multiple';
-
-  return (
-    <div>
-      <label className="text-sm">Group</label>
-      <Menu as="div" className="relative text-left">
-        <div>
-          <Menu.Button className="relative w-full cursor-default rounded-md border bg-white py-2 pl-3 pr-10 text-left text-sm shadow-sm focus:border-indigo-500 focus:outline-none focus:ring-1 focus:ring-indigo-500 dark:border-gray-600 dark:bg-gray-900 sm:text-sm">
-            <span className="block overflow-x-hidden text-ellipsis">{label}</span>
-            <span className="pointer-events-none absolute inset-y-0 right-0 ml-3 flex items-center pr-2 text-gray-400">
-              <Icon icon="heroicons:chevron-down-20-solid" aria-hidden="true" />
-            </span>
-          </Menu.Button>
-        </div>
-
-        <Transition
-          as={Fragment}
-          leave="transition ease-in duration-100"
-          leaveFrom="opacity-100"
-          leaveTo="opacity-0"
-        >
-          <Menu.Items className="absolute left-0 z-10 mt-1 min-w-[400px] overflow-auto rounded-md bg-gray-50 py-1 text-base shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none dark:border-gray-600 dark:bg-gray-900 dark:ring-white dark:ring-opacity-20 sm:text-sm">
-            <div className="p-4">
-              <fieldset>
-                <div className="space-y-5">
-                  {groupByOptions.map(({value, label, description, disabled}) => (
-                    <div key={value} className="relative flex items-start">
-                      <div className="flex h-6 items-center">
-                        <input
-                          id={value}
-                          name={value}
-                          type="checkbox"
-                          disabled={disabled}
-                          className="h-4 w-4 rounded border-gray-300 text-indigo-600 focus:ring-indigo-600"
-                          checked={groupBy.includes(value)}
-                          onChange={() => toggleGroupBy(value)}
-                        />
-                      </div>
-                      <div className="ml-3 text-sm leading-6">
-                        <label
-                          htmlFor={value}
-                          className="font-medium text-gray-900 dark:text-gray-200"
-                        >
-                          {label}
-                        </label>
-                        <p className="text-gray-500 dark:text-gray-400">{description}</p>
-                      </div>
-                    </div>
-                  ))}
-                </div>
-              </fieldset>
-            </div>
-          </Menu.Items>
-        </Transition>
-      </Menu>
-    </div>
-  );
-};
-
-const SortBySelect = ({
-  sortBy,
-  setSortBy,
-  compareMode,
-}: {
-  sortBy: string;
-  setSortBy: (key: string) => void;
-  compareMode: boolean;
-}): React.JSX.Element => {
-  return (
-    <div>
-      <label className="text-sm">Sort</label>
-      <Select
-        items={[
-          {
-            key: FIELD_FUNCTION_NAME,
-            disabled: false,
-            element: {
-              active: <>Function</>,
-              expanded: (
-                <>
-                  <span>Function</span>
-                </>
-              ),
-            },
-          },
-          {
-            key: FIELD_CUMULATIVE,
-            disabled: false,
-            element: {
-              active: <>Cumulative</>,
-              expanded: (
-                <>
-                  <span>Cumulative</span>
-                </>
-              ),
-            },
-          },
-          {
-            key: FIELD_DIFF,
-            disabled: !compareMode,
-            element: {
-              active: <>Diff</>,
-              expanded: (
-                <>
-                  <span>Diff</span>
-                </>
-              ),
-            },
-          },
-        ]}
-        selectedKey={sortBy}
-        onSelection={key => setSortBy(key)}
-        placeholder={'Sort By'}
-        primary={false}
-        disabled={false}
-      />
-    </div>
   );
 };
