+++ ui/packages/shared/profile/src/ProfileView/VisualizationPanel.tsx
@@ -51,8 +51,15 @@ export const VisualizationPanel = React.memo(function VisualizationPanel({
 
   return (
     <>
-      <div className="flex w-full items-center justify-end gap-2 pb-2">
-        <div className="flex w-full items-center justify-between flex-col-reverse md:flex-row">
+      <div className="flex w-full items-center justify-end gap-2 pb-2 min-h-[78px]">
+        <div
+          className={cx(
+            'flex w-full justify-between flex-col-reverse md:flex-row',
+            isMultiPanelView && dashboardItem === 'icicle'
+              ? 'overflow-x-scroll items-end gap-x-2'
+              : 'items-center'
+          )}
+        >
           <div className="flex items-center">
             <div
               className={cx(isMultiPanelView ? '' : 'hidden', 'flex items-center')}
@@ -62,7 +69,12 @@ export const VisualizationPanel = React.memo(function VisualizationPanel({
             </div>
             <div className="flex gap-2">{actionButtons}</div>
           </div>
-          <div className="flex flex-col items-end gap-4">
+          <div
+            className={cx(
+              'flex flex-col items-end gap-4',
+              isMultiPanelView && dashboardItem === 'icicle' && 'pb-2'
+            )}
+          >
             <ViewSelector defaultValue={dashboardItem} navigateTo={navigateTo} position={index} />
 
             {dashboardItem === 'icicle' && flamegraphHint != null ? (
