+++ ui/packages/shared/profile/src/ProfileView/index.tsx
@@ -37,9 +37,7 @@ import {
 import {
   Button,
   ConditionalWrapper,
-  IS,
   KeyDownProvider,
-  TS,
   UserPreferences,
   useParcaContext,
   useURLState,
